requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fastapi==0.115.12
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.53.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
gluonts==0.16.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.8.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.37.1
numpy==1.26.4
openai==1.76.2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.29.3
pyarrow==20.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit>=1.30.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.2
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==2.1.4
